api/requirements.txt
fastapi==0.115.8
fastapi-cli==0.0.7
requests==2.32.3
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.7.1
pydantic_core==2.27.2
passlib==1.7.4
SQLAlchemy==2.0.38
redis==5.2.1
python-dotenv==1.0.1
python-jose==3.4.0
passlib==1.7.4
psycopg2==2.9.10
dnspython==2.7.0
email-validator==2.2.0


